Add files via upload
TrigEdit trigger list now uses normal text colors (had to find the constants). If a trigger's first condition or first action (preferring the former) is a comment with non-empty text, that comment shows as a description for the trigger in the trigger list. Triggers now have an "enabled" flag and a list of entry points. Trigger serialization has been changed such that triggers contain "data chunks," with the condition and action lists being chunks; this change makes room in the format for an arbitrary number of trigger fields; currently the "enabled" field is also serialized and parsed. The TrigEdit trigger list now shows a checkbox on each trigger, which is intended to be used to enable and disable triggers individually (such that you don't have to delete a trigger wholesale to turn it off).