TrigEdit
Added "Move Up" and "Move Down" buttons to the editor (not yet functional). The "Cancel" button in the trigger editor is now functional; the "New" button in the trigger list is now functional. Fixed sizing for the trigger editor's opcode lists (which formerly overlapped the "Cancel" button, blocking clicks to it). Opcode lists in the trigger editor now use a template. You can now double-click a trigger in the trigger list to edit it.