Add files via upload
Trigger opcode descriptions now use Zenimax format codes, allowing arguments to be visually reordered. Opcode argument color is consistent between TriggerEdit and OpcodeEdit. Fixed a copy-and-paste mistake that caused OpcodeEdit to fail to open. OpcodeEdit now opens in a manner fully consistent with OpcodeArgEdit, and TriggerEdit redraws a trigger's opcodes after changes are made. There is now code to allow you to change an opcode's type in OpcodeEdit; the opcode's arguments are reset to appropriate defaults for the new OpcodeBase.