Add files via upload
Started work on refactoring the "widget" classes. Added a single "WWindow" class that unifies the functionality of WModalHost and WModal, while also offering standard title bars for windows; it is currently used by two test menus, and the plan is to take the existing TrigEdit windows and refactor them as singletons that subclass WWindow. Deferreds now have an "always" method to attach callbacks that fire on resolve or reject. Deferreds now clear their callback lists after callbacks are executed on resolve or on reject. Added a helper file with benchmarking functions (not yet tested).