Add files via upload
Fixed copy/paste mistakes in Deferred. You can now create a read-only proxy to a Deferred, known as a Promise. Removed underscore prefixes from most of the WWindow methods that are meant to be overridden by subclasses, and mass-updated all subclasses accordingly; created copies of some methods in order to remove the need to call super in overrides. OpcodeArgEdit now shows a confirmation prompt when you cancel, but only if you're editing an argument with data that is considered "effortful" (currently any string argument).