Theming
More work on theming, and some fixes and changes to functionality in service to theming: almost all theme colors are now defined in a single Lua file, to allow for rapid swapping/testing of new themes; custom buttons; a WScrollSelectList template with theming code; added WScrollList:onRedrawOrReposition() to help theming code. Made a fix to WidgetClass: trying to install a widget on the same control twice will return the existing widget on the second call (or, if a widget of a different class but the same key name is there, an assertion will fail).