WCombobox
Started work on a custom combobox control, using some of the vanilla combobox code as a reference but attempting to base mine on WScrollSelectList. The class is still incomplete; among other things, click handlers are generally broken, theming code is unfinished, mouseover behaviors are unfinished, and there are faults stemming from design flaws in WScrollList (namely: if the dropdown list hasn't been drawn yet, then selection cannot be influenced programmatically; if the drawn controls are "out of date," then programmatic influence will produce incorrect results; and this applies to all WScrollLists, not just the one in the combobox).