WCombobox
WCombobox restored to full functionality, and fully themed. All existing windows modified to use WCombobox instead of the Zenimax combobox. Added a "focus ring" that shows when the combobox is opened, to visually separate the selected item from the dropdown body. Added a border to the dropdown body to differentiate it from any other white controls that may be behind it.