WScrollList and WCombobox fixes
WCombobox now closes if you click the already-selected dropdown item. WScrollList contains workarounds for an issue in the UI system: OnMouseEnter and OnMouseExit fire if the mouse actively moves over or off of a control, but not if the control itself is programmatically moved under or out from under the mouse, and this negatively impacts WCombobox when using the mouse wheel to scroll the dropdown; WScrollList now checks, after every scroll operation, what the current mouseover control is, and manually fires OnMouseEnter and OnMouseExit when and as appropriate.