Savedata
Triggers are now saved per-character. Created a custom interface for savedata comparable to ZO_SavedVars; this interface allows us to update out-of-date savedata instead of uncontrollably nuking it out of existence. The trigger list window now saves all triggers (i.e. prompts them to serialize) whenever it's hidden. Fixed a bug in WScrollList where removing a single list item wouldn't cause the list to redraw properly (the cleanup loop, meant to deal with now-unused elements, was off-by-one). WWindow's padding is no longer scaled to counteract ESO's resolution normalization (there are anchors all over the UI that are meant to be consistent with the padding, but I'm not sure that they can use arbitrary Lua variables or expressions -- something to revisit -- so for now, we're just using a constant number). Added an untested function to the table utils for detecting cyclical references in a table.