Add files via upload
List items in the trigger list and opcode lists now word-wrap properly. Set specific maximum character counts on string and multi-line values in the opcode argument editor. WScrollList's redraw method now redraws all list items twice in a row, in order to address issues with UI reflow (which came up when adding the word-wrap fixes); this is an absolutely hideous workaround, but it's the best we can do without any easily-available documentation of the UI engine internals.