Nested trigger UI support
The trigger editor can now edit nested triggers: they are clickable in the opcode editor, and the trigger editor supports a "stack" of editor states to allow for editing. UI flow is as follows: click the argument when editing a nested trigger opcode; the opcode window will close and you will be returned to the trigger edit window, which will now display the nested trigger; clicking "OK" or "Cancel" will make your changes and return you to the parent trigger. The trigger list window now supports resizing as well, which will be vital when we work on a way to display the contents of nested triggers.