Add files via upload
WBulletedList can now limit the depth to which it displays list items, with three options: don't show elements past a specified nesting level; stop showing elements once we've indented past half of the container's width; or stop showing elements once we have less than a specified amount of space remaining due to indentation. List items that are disqualified by these criteria are skipped when rendering, but there is also an option to show a placeholder string (e.g. "...") in place of the first list item to disqualify in a row.

In addition to this, a critical fix was made to the parsing of serialized triggers.