Add files via upload
Added a condition to check an item's base type (itemType). Fixed the handling of enums with non-contiguous keys in various parts of the mod, including the opcode argument editor. Triggers no longer execute if their "enabled" flag is false. Fixed the serialization and parsing of special characters in trigger data. Added a trigger action to destroy the item stack being processed. Modified the "log message" action: the substring "$(name)" will now be replaced with the name of the item stack being processed.