Add files via upload
Fixed a bug that made it impossible to properly edit an opcode if its description word-wrapped within the window; on a separate but related note, the opcode editor now resizes to fit the description. Added a condition to check the cause of an item being added to the player's inventory; this condition relies on state-tracking based on the detection of open/close events for mail, merchants, and banks, since the "item slot updated" event itself doesn't tell us anything useful. Added a condition to tell whether an item was crafted by a player. Benchmarked the ItemInterface.