Add files via upload
Fixed a bug in the trigger editor and opcode editor windows: ESO chokes on formatting codes that apply to strings that are too long: the game truncates affected strings and displays them improperly, often showing only a few characters or even failing to show any text at all; affected strings are truncated to 253 characters, but the limit itself seems to be lower than that. Format operations now split all strings at 200-character boundaries and format each chunk individually. In some cases, a string may be split at a space, and the game will decline to format that space; however, the space still shows, so I... don't care. I don't care. It's fine. It's fine! It's fine.