Add files via upload
Added an "Added Item Is New Stack" condition; this should make it possible for triggers to be "considerate" when destroying items even without the ability to destroy just part of a stack: the trigger can just only destroy new stacks. Fixed some bugs related to the "Set And/Or" condition: we now reset the "matched or" flag when switching to using "or," and we now properly handle "or" conditions at the end of a trigger's condition list. Added trigger actions, untested, for modifying an item's "junk" flag and for laundering an item.