Add files via upload
Implemented the ability to destroy just part of a stack, and restored the "Destroy" trigger action to allow that as an option. Added a condition to query the total count of an item in the inventory, bank, or crafting bag. Opcode failures are now dumped to the chatbox. "Extra" parameters passed through the ConditionBase and ActionBase constructors now reach OpcodeBase properly. Trigger:exec now enforces opcodes' entry point limitations, and now properly returns extra error information when an opcode fails.