Add files via upload
Created a system for splitting item stacks; this should prevent bag slot conflicts when multiple stacks are split in a short period of time. Modified how Deferred callbacks work: they are no longer passed the Deferred instance as an argument, and a callback registered without a "context" will receive only the arguments that the Deferred was resolved or rejected with (instead of a nil first argument).