Localization
Created custom strings for the different kinds of raw and refined materials (clothing, jewelry, smithing, woodworking) in the itemType enum. This should solve the problem where all of the former were localized as indistinguishable "Raw Material" and all of the latter were localized as indistinguishable "Material," which should in turn make the Item Type trigger condition marginally easier to use.