Add files via upload
Improved how the "log message" action handles tokens. Added conditions for whether an item is researchable and for who it was crafted by (allowing "$(player)" as a special token). Fixed existing code for deconstructing an item, and added a "deconstruct" trigger action. Fixed a bad variable name in the "Sell or Fence" trigger action. Replaced several table.getn calls with the modern Lua "#" operator, and improved the functionality of the assign and assignDeep utility functions.