Add files via upload
Improved the trigger condition for testing an item's style. Discovered that we can't respond to bank withdrawals without also responding to the splitting of stacks within the same bag (which we really, really, really, really, really don't want to do); it has been made possible to disable specific keys within an opcode's enum argument, and the "was withdrawn from the bank" option in the "Added Item Cause" condition has been dummied out in this manner.