LibAddonMenu integration and more
Integrated LibAddonMenu-2.0 and added a prefs menu. Added the ability to log trigger failures and misses. Added a condition to trigger "miss" logging. Added a trigger to the (currently inaccessible) trigger gallery, for auto-discarding low-rarity stolen items unless they can be stockpiled for The Covetous Countess. Fixed OR-condition short-circuiting: opcodes now have a way to define that they shouldn't be skipped. Added prefs to control whether to skip locked, Crown Crate, and Crown Store items (the latter two are untested); a pref to control whether to pre-sort items before running fence triggers on them; and prefs to control logging.