Autocomplete; Halt Trigger action
Made it possible for opcodes' string arguments to have autocomplete. Added an enum for commonly-used items, and set it up as the autocomplete entry list for the Item Name condition. Added a pref to globally enable/disable autocomplete; by default, autocomplete is disabled because we can't theme it (it uses ZO_AutoComplete, which itself relies on a hardcoded(?) menu widget). Added a trigger action that halts all trigger processing on the current item; this can be useful for letting the player control what triggers are allowed to run on (e.g. if you don't want to run triggers on fishing lures, then the first trigger in your list, enabled for all entry points, can be one that halts processing if the item is a lure).