WCombobox autocomplete
Added autocomplete to WCombobox: you can now type within a combobox in order to jump to an entry by name. This relies on an internal invisible EditBox, which we also piggyback off of to handle Escape, Enter, and the up and down arrow keys. Fixed lazy scrolling-to-index for WScrollList when scrolling down.