Pretend Mode; fence fixes; gallery
Triggers in the trigger gallery no longer use the Log Message action to log what they do, since this is covered by the userpref to log all actions taken on an item. Added a trigger to the trigger gallery for destroying various stolen items that a player would commonly consider unimportant. Fixed the behavior of quantity-enum options in the argument editor, following a refactor to that window. Added trigger conditions for checking the current launder and fence transaction limits. The trigger actions for laundering and fencing now return failure codes if the daily transaction limits are hit, after laundering/fencing as much of the stack as is possible. Added a "Pretend Mode" option, which makes specific trigger actions (anything that would influence an item) simply log what they *would* do, and otherwise do nothing.