Add files via upload
"Quantity" opcode arguments now use a class, OpcodeQuantityArg, which centralizes most related functionality. "Quantity" and "number" arguments now have fields that can be used to validate player-entered values. Trigger:exec now validates an opcode's arguments before executing it; an opcode that fails to value will cause the containing trigger to fail. Modified how triggers are run: if a trigger fails to run because an opcode had invalid arguments, or because an opcode was not valid for the current entry point, then the trigger will not be executed for the remainder of an entry point (i.e. if we're running triggers on every item in the inventory, the faulting trigger will be skipped for the remaining items). Several trigger conditions and actions now use the new validation fields in their arguments. Some minor miscellaneous fixes.