Add files via upload
Fixed the lazy getters on ItemInterface. Added a lazy getter for an item's treasure categories. Added conditions to check an item's treasure categories, to check if an item is a known recipe/pattern/motif, and to check how the item relates to the Covetous Countess quest. Created QuestInterface. Modified the opcode argument editor's autoSize function in an attempt to avoid especially narrow windows.