Add files via upload
Added a user pref to control whether Esc force-closes the editor. Created the WTooltip widget, useful for showing a text-only tooltip on either side of some target (you can specify an axis and a side to prefer). The opcode argument editor now has a WTooltip; if the value you've typed is invalid, then a tooltip will be shown adjacent (vertically) to the window, stating the correct range of values.