Add files via upload
Added a condition to check whether your crafting skills (specific ones, or any, or all, or the specific one related to the current item) are maxed. Opcode-level explanation text is now shown in the opcode editor. Added explanation text for "Set And/Or," and clarified explanation text for "Log Trigger Miss." The "Can Launder" and "Can Fence" conditions are no longer restricted to the "Fence Menu" entry point.