Theming system and dark theme
Added a theme switcher, exposed via the LibAddonMenu options menu, and created a dark theme that uses ESO's native color palette for PC. Added some new triggers to the trigger gallery, and reordered them a bit as well. Fixed existing trigger gallery triggers to honor an item's "locked" flag. Fixed a bug with the "Esc force-closes editor" option that caused the default action for the Esc key to always be suppressed. Fixed spacing between the opcode argument editor and its validation tooltip. Fixed a bad frame handler in the trigger edit window: there was a check to prevent it from running every frame (it was only meant for resize), but the check was busted.