v1.0.6
Added two new entry points that will run triggers on items in your bank: "Bank Opened (run on banked items)" and "Crafting Menu Opened (run on banked items)." The intended use case for this is being able to automatically deconstruct banked items.

The "Deposit in Bank" action will now cause a trigger failure and show an error message when attempting to deposit Character Bound items, instead of just failing silently.

Added a "Bound" condition to test whether an item is bound, or to test whether an item is Character Bound, specifically.

Zenimax forcibly disconnects players from the server if their add-ons attempt to deposit more than 99 item stacks in the bank during one bank session. ItemTrig is now aware of this and will attempt to stay 2 stacks below the limit. Note that ItemTrig cannot be aware of what other add-ons are doing, so ItemTrig and other add-ons can still exceed the limit together (this is also the case for laundering items).

Fixed a bug that could cause attempts to enable/disable a top-level trigger to fail.

Added an option to improve logging of trigger failures by "collapsing" certain kinds of trigger failures when they occur more than once. For example, if you have a trigger that tries to deposit ten items into your bank, and your bank is full, you do not need to see ten error messages all telling you that the bank is full. This option will instead display full details on the first such trigger failure, and then (after all triggers for the entry point have finished running) display the number of triggers that experienced the same error. This option is enabled by default.

Added an option to forcibly prevent triggers from destroying items in your bank, even if they run on bank-specific entry points. If the option is enabled, then Destroy actions on a banked item will cause a trigger failure. This does not apply to deconstructing items.

Fixed WTooltip's positioning code, and added a WTooltipInPlace subclass that does Windows-style in-place tooltips.